requirements.txt
#
# This file is autogenerated by pip-compile with python 3.8
# To update, run:
#
#    pip-compile requirements.in
#
distro==1.7.0
    # via tabula-py
numpy==1.23.2
    # via
    #   pandas
    #   tabula-py
pandas==1.4.4
    # via
    #   -r requirements.in
    #   tabula-py
python-dateutil==2.8.2
    # via pandas
pytz==2022.2.1
    # via pandas
six==1.16.0
    # via python-dateutil
tabula-py==2.5.1
    # via -r requirements.in
